RF: Centralised theme spec
